Book 3: probabilistic sampling (#5)
This PR implements the features from the third book in https://raytracing.github.io/.

There are still a lot of improvements to be made, bunch of code that needs to be cleaned up, and even some features missing from certain places (e.g. some functions are not implemented to all object types yet!)

However, merging this in order to have one more thing "done" and to make it easier to scope next changes into smaller PRs too.